allow optionally pasting above/below if requested
... by passing an optional `"p"` or `"P"` argument to `peekup_open`.

Or, well, things work only if one passes `nil`, `""`, `"p"` or `"P"` to
it. YMMV.

If one passes `""` or `nil`, they get the same behaviour as before - the
contents of the `*` register are replaced with the contents of the
picked register.

If one passes `"p"` or `"P"` by creating additional mappings like:

    nnoremap <Plug>PeekupOpenThenPasteAfter  :lua require('nvim-peekup').peekup_open("p")<CR>
    nnoremap <Plug>PeekupOpenThenPasteBefore :lua require('nvim-peekup').peekup_open("P")<CR>

... the contents of the `*` register are *not* replaced, but instead the
contents of the _picked_ register are pasted _before_ or _after_ where
the cursor was prior to invoking the `""` mapping, by "simply" pasting
the register before or after.

If one passes `"P"` to the function, the contents of the picked register
are, instead, pasted above.

Example suggested mappings:

    nmap <Leader>P <Plug>PeekupOpenThenPasteBefore
    nmap <Leader>p <Plug>PeekupOpenThenPasteAfter